requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

keras==3.3.2
pillow==10.3.0
